+++ crates/amaru-pure-stage/src/stagegraph.rs
@@ -121,13 +121,6 @@ pub trait StageGraph {
         Msg: SendData + serde::de::DeserializeOwned,
         St: SendData;
 
-    fn contramap<Original: SendData, Mapped: SendData>(
-        &mut self,
-        stage_ref: impl AsRef<StageRef<Original>>,
-        new_name: impl AsRef<str>,
-        transform: impl Fn(Mapped) -> Original + 'static + Send,
-    ) -> StageRef<Mapped>;
-
     /// Preload the given stage’s mailbox with the given messages.
     ///
     /// Since the stage is not running yet, this will return false and drop messages